+++ src/data/caseStudy.js
@@ -1,0 +1,22 @@
+const caseStudy = {
+  image: 'https://images.unsplash.com/photo-1461749280684-dccba630e2f6?auto=format&fit=crop&w=600&q=80',
+  title: 'Enterprise IT Modernization for Fortune 500 Company',
+  challenge: 'Modernize legacy IT across 10 countries with reduced downtime.',
+  solution: [
+    'Migrated systems to hybrid cloud',
+    'Deployed monitoring tools for real-time performance',
+  ],
+  results: [
+    '40% reduction in operational costs',
+    '75% improvement in deployment speed',
+    '99.9% uptime achieved across global offices',
+  ],
+  person: {
+    name: 'Anand Iyer',
+    role: 'Global IT Director',
+    avatar: 'https://randomuser.me/api/portraits/men/32.jpg',
+  },
+  quote: 'Zero Innovation helped us scale globally with a future-proof IT strategy. Their technical expertise and execution were flawless.',
+};
+
+export default caseStudy;
