+++ src/data/heroMetrics.js
@@ -1,0 +1,8 @@
+const heroMetrics = [
+  { value: '500+', label: 'Projects Completed' },
+  { value: '98%', label: 'Client Satisfaction' },
+  { value: '25+', label: 'Industries Served' },
+  { value: '15', label: 'Years Experience' },
+];
+
+export default heroMetrics;
