+++ src/data/metrics.js
@@ -1,0 +1,8 @@
+const metrics = [
+  { icon: '/Overlay-1.svg', value: '100+', label: 'Products Deployed' },
+  { icon: '/Overlay-4.svg', value: '95%', label: 'Client Retention Rate' },
+  { icon: '/Overlay-3.svg', value: '15%', label: 'Average Annual Returns' },
+  { icon: '/Overlay-2.svg', value: '4 Weeks', label: 'Average Implementation Time' },
+];
+
+export default metrics;
