+++ src/data/navItems.js
@@ -1,0 +1,10 @@
+const navItems = [
+  { name: 'Home', href: '#' },
+  { name: 'Services', href: '#' },
+  { name: 'About', href: '#' },
+  { name: 'Projects', href: '#', active: true },
+  { name: 'Blog', href: '#' },
+  { name: 'Contact', href: '#' },
+];
+
+export default navItems;
