+++ src/data/projects.js
@@ -1,0 +1,49 @@
+const caseStudyImage = 'https://images.unsplash.com/photo-1461749280684-dccba630e2f6?auto=format&fit=crop&w=600&q=80';
+
+const projects = [
+  {
+    category: 'AI Applications',
+    title: 'AI-Powered Healthcare Assistant',
+    description: 'Custom-built NLP-based virtual assistant for a 400-bed hospital, improving patient interaction efficiency by 65%.',
+    image: caseStudyImage,
+  },
+  {
+    category: 'SaaS Platforms',
+    title: 'SaaS CRM for Real Estate Firm',
+    description: 'Developed a responsive SaaS CRM, enabling 3x faster deal closures.',
+    image: caseStudyImage,
+  },
+  {
+    category: 'Web Development',
+    title: 'E-learning Platform',
+    description: 'Built scalable LMS supporting 50K+ users with analytics and payments.',
+    image: caseStudyImage,
+  },
+  {
+    category: 'Cloud & DevOps',
+    title: 'IT Infrastructure Revamp',
+    description: 'Migrated to cloud, improved efficiency, reduced costs.',
+    image: caseStudyImage,
+  },
+  {
+    category: 'AI Applications',
+    title: 'AI-based Resume Screening Tool',
+    description: 'Scans 10K+ resumes/day using skill matching and NLP.',
+    image: caseStudyImage,
+  },
+  {
+    category: 'Web Development',
+    title: 'EdTech Platform for Teachers',
+    description: 'Training portal for 200+ educators, with certification tracking.',
+    image: caseStudyImage,
+  },
+];
+
+export const projectTabs = [
+  'All Projects',
+  'AI Applications',
+  'SaaS Platforms',
+  'Web Development',
+];
+
+export default projects;
